Modified how the read status is updated to use the new email schema.  The email id must be specified in the query string to locate the email.
The image to return must be specified in the query string and must be located in the `public/modules/core/img/email` directory.  If this field is not specified, the logo.png image will be returned.

Errors are no longer returned (unless the server crashes).  If a field is not specified, either the specified image is returned or the default image is returned.